refactor(infra): share guarded body-stream cleanup across fetch consumers
